% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{EffDist}
\alias{EffDist}
\title{Default distributions of fishing effort, selectivity and catches}
\description{
Default distributions of fishing effort, selectivity and catches
}
\author{
Tom Carruthers \email{tom@bluematterscience.com}
}
\keyword{data}
