% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_tools.R
\name{fakeMP}
\alias{fakeMP}
\title{An MP that stores a data object while a projection is running}
\usage{
fakeMP(
  Data,
  repyr = 2023,
  outdir = "C:/GitHub/DolphinMSE/MP_design",
  filenam = "Example_Data",
  Eff = 0.5
)
}
\arguments{
\item{Data}{An MSEtool object of class 'Data'}

\item{repyr}{The projected time step in which the data is written to disk. Positive integer.}

\item{outdir}{A character string representing the file directory to save Example_Data to.}

\item{filenam}{A character string representing the file name}

\item{Eff}{Real number, the constant effort for projection}
}
\description{
An MP that stores a data object while a projection is running
}
\author{
T. Carruthers
}
