% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_tools.R
\name{get_season}
\alias{get_season}
\title{MP helper function - what season is it?}
\usage{
get_season(Data)
}
\arguments{
\item{Data}{An MSEtool object of class 'Data'}
}
\description{
MP helper function - what season is it?
}
\author{
T. Carruthers
}
