% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahiMP.R
\name{mahiMP}
\alias{mahiMP}
\title{The Mahi Management Procedure (OpenMSE v2.0)}
\usage{
mahiMP(
  Data,
  TAC = NA,
  rel_TAC = NA,
  TAE = NA,
  rel_TAE = 1,
  SLmin = c(0, 0, 500, 0, 500, 0, 0, 0),
  SLmax = c(10000, 10000, 10000, 10000, 10000, 10000, 10000, 10000),
  TL = c(NA, 54, 54, 54, 54, NA, NA, NA),
  PRM = list(0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25),
  HCR_use = F,
  HCR_ICP = NA,
  HCR_LCP = NA,
  HCR_calib = 145:148,
  HCR_fleet = list(NA),
  HCR_sens = 1,
  HCR_lever = "TAC",
  HCR_up_max = 0.5,
  HCR_down_max = 0.5,
  HCR_input = 1,
  HCR_smooth = NA,
  Index_type = "VAST",
  Index_cv = NA,
  Index_ac = NA,
  Index_lag = 1,
  Index_seed = 1,
  TL_empirical = T,
  plot = F,
  ...
)
}
\arguments{
\item{Data}{An object of openMSE class 'Data'}

\item{TAC}{Annual total allowable catch. A single value (all fleets, all areas), a vector length nFleet or a matrix nFleet x nArea. Positive real number.}

\item{rel_TAC}{A single value that is current TAC relative to the most recent  historical year. Positive real number. Can be a single number, a vector nfleet long or a matrix nfleet by narea long.}

\item{TAE}{Annual total allowable effort. Single value (all fleets, all areas) or a vector length nFleet. Positive real number.}

\item{rel_TAE}{Relative effort (an improper fraction of effort in the last historical year). Positive real number. Single number (all fleets/areas), vector of length nfleets (8), vector of length nareas, or matrix with nfleets rows and nareas columns (8 x 7).}

\item{SLmin}{Minimum size limit (mm). Positive real number. Can be a single number (all fleets) or a vector nfleets long (8).}

\item{SLmax}{Maximum size limit (mm). Positive real number larger than Smin (if specified). Can be a single number (all fleets) or a vector nfleets long (8).}

\item{TL}{Trip limit (fish retained per trip). Positive integer. Can be a single number (all fleets) or a vector nfleets long (8). Note that predictive models are only available for RecS and HireS in regions CAR+FLK and NCFL.}

\item{PRM}{Post Release Mortality (discard mortality). List nfleet long of either a Fraction (all length classes) or a vector nlengths long. Default from US hook and line estimates of Rudershausen et al 2019 (DOI: 10.1002/nafm.10348). Can be a single number or a vector of length nareas (7).}

\item{HCR_use}{Should HCR(s) be used? Boolean. If False, all other HCR arguments will be ignored.}

\item{HCR_ICP}{Index control points. These are a pair of x-axis control points for a hockey-stick control rule. Phrased in terms of the levels of the index (HCR_input) over the HCR_calib years.}

\item{HCR_LCP}{Management Lever control points. These are a pair of y-axis control points for a hockey-stick control rule. Phrased in terms of the levels of the Lever (HCR_lever) over the HCR_calib years.}

\item{HCR_calib}{The calibration years used - which historical time period is used? Vector of integer time steps.}

\item{HCR_fleet}{The fleet for which the HCR is used. Vector of positive integers nHCR long. If NA this sets advice for all fleets combined based on the HCR_input specified.}

\item{HCR_sens}{Sensitivity of the response. A vector of positive real numbers nHCR long. A value of 1 is proportional.}

\item{HCR_lever}{The output variable. A vector of character strings nHCR long. E.g., "TAC", "TAE".}

\item{HCR_up_max}{The maximum rate of advice increase. Positive imperfect fraction. E.g., 0.2 is a maximum management increase of +20 per cent.}

\item{HCR_down_max}{The maximum rate of advice decrease. Positive fraction. E.g. 0.8 is a maximum management decreate of 80 per cent.}

\item{HCR_input}{The indices used in the harvest control rule. A list of integer vectors, nHCR long. E.g., HCR_input = list(2:3, c(1,3)) uses indices 1 and 2 for the first HCR, indices 1 and 3 for the second HCR}

\item{HCR_smooth}{The effective number of parameters of a polynomial smoother applied to the input indices. Defaults to NA - raw data. A value of 0.1 means that there will be length(Index)*0.1 number of smoothing parameters. Hence, higher values are less smoothing. This parameterization keeps smoothing consistent as the length of the time series increases in projections.}

\item{Index_type}{The origin of the index. Character string nIndex long. Can be biomass ("B"), spawning stock biomass ("SSB"), vulnerable biomass of fleet 1 ("VB1"), vulnerable biomass of fleet 2 ("VB2"), etc.}

\item{Index_cv}{The coefficient of variation of the index (CV). Vector of positive real numbers nIndex long.}

\item{Index_ac}{The lag-1 autocorrelation of the index. (Generally positive) Vector of fractions nIndex long.}

\item{Index_lag}{The number of years of data lag. Vector of integers, nIndex long. 1 is a single year.}

\item{Index_seed}{Real number, a random number seed for generating index observation errors nIndex long}

\item{TL_empirical}{Should the empirical distribution of catch rates be used to calculate release rate according to trip limits?}

\item{plot}{Should plots be produced that explain MP calculations (e.g. HCRs)? Boolean.}
}
\description{
A function for emulating status quo mahi management and allowing for a wide range of combined management controls. Fleets (n=8) are: (1) USCom, (2) RecN, (3) RecS, (4) HireN, (5) HireS, (6) Intl, (7) Disc, (8) UnRep. Areas (n = 5) are:  (1) "CAR+FLK", (2) "NCA", (3) "NCFL", (4) "NED", (5) "NNC+VBM"
}
\examples{
mahiMP(Example_Data)
}
\author{
T. Carruthers
}
