% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahiplot.R
\name{mahiplot}
\alias{mahiplot}
\title{Plot overall yield and biomass projections across MPs}
\usage{
mahiplot(
  MSEobj,
  MPs = 1:2,
  annual = T,
  ns = 4,
  stock = 1,
  MPcols = c("#0000ff50", "#ff000050", "#00ff0050"),
  npy = 20,
  MPnams = NA,
  leg = T,
  addn = F,
  addmed = T,
  addl = F
)
}
\arguments{
\item{MSEobj}{A class of object 'MSE'}

\item{MPs}{An integer vector of MPs to plot}

\item{annual}{Boolean, should annual or seasonal results be plotted?}

\item{ns}{Integer, number of seasons.}

\item{MPcols}{Character vector of colors for plotting MP projections}

\item{npy}{Positive integer, number of plot years - No. historical years to plot before projection}

\item{leg}{Boolean, add a legend for the MPs?}

\item{addn}{Boolean, add narrow interquartile range shading?}

\item{addmed}{Boolean, add a line for the median?}

\item{addl}{Boolean, add individual simulation lines?}

\item{MPnam}{Character vector, optional, a character vector of MP names}
}
\description{
Plot overall yield and biomass projections across MPs
}
\examples{
mahiplot(myMSE)
}
\author{
T. Carruthers
}
