% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahiplot2.R, R/mahiplot_MSY.R
\name{mahiplot2}
\alias{mahiplot2}
\title{Plots projected exploitation rate and spawning biomass relative to MSY  across MPs}
\usage{
mahiplot2(
  MSElist,
  MPs = 1:2,
  annual = T,
  ns = 4,
  stock = 1,
  MPcols = c("#0000ff50", "#ff000050", "#00ff0050"),
  npy = 20,
  MPnams = NA,
  leg = T,
  addn = F,
  addmed = T,
  addl = F
)

mahiplot2(
  MSElist,
  MPs = 1:2,
  annual = T,
  ns = 4,
  stock = 1,
  MPcols = c("#0000ff50", "#ff000050", "#00ff0050"),
  npy = 20,
  MPnams = NA,
  leg = T,
  addn = F,
  addmed = T,
  addl = F
)
}
\arguments{
\item{MSElist}{A list of objects of class 'MSE' - must be at least 2 MSE objects long}

\item{MPs}{An integer vector of MPs to plot}

\item{annual}{Boolean, should annual or seasonal results be plotted?}

\item{ns}{Integer, number of seasons.}

\item{MPcols}{Character vector of colors for plotting MP projections}

\item{npy}{Positive integer, number of plot years - No. historical years to plot before projection}

\item{leg}{Boolean, add a legend for the MPs?}

\item{addn}{Boolean, add narrow interquartile range shading?}

\item{addmed}{Boolean, add a line for the median?}

\item{addl}{Boolean, add individual simulation lines?}

\item{firstyr}{Calendar year, first year to plot results}

\item{lastyr}{Calendar year, last year to plot results}

\item{MSEobj}{A class of object 'MSE'}

\item{MPnam}{Character vector, optional, a character vector of MP names}
}
\description{
Plots projected exploitation rate and spawning biomass relative to MSY  across MPs

Plots projected exploitation rate and spawning biomass relative to MSY  across MPs
}
\examples{
mahiplot2(list(myMSE, myMSE))
mahiplot2(myMSE)
}
\author{
T. Carruthers
}
