% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahiMP.R
\name{mahiMP}
\alias{mahiMP}
\title{The Mahi Management Procedure}
\usage{
mahiMP(
  x,
  DataList,
  Effort = NA,
  TAC = NA,
  Smin = NA,
  Smax = NA,
  BL = NA,
  PRM = 0.25,
  ...
)
}
\arguments{
\item{x}{Simulation number}

\item{DataList}{A list object [[nstocks]][[nareas/fleets]] of openMSE objects of class 'Data'}

\item{Effort}{Relative effort (an improper fraction of effort in the last historical year). Positive real number. Single number (all fleets/areas), vector of length nfleets (8), vector of length nareas, or matrix with nfleets rows and nareas columns (8 x 7).}

\item{TAC}{Annual TAC (kg). Positive real number. Single number (all fleets/areas), vector of length nfleets (8), vector of length nareas (7), matrix with nfleets rows and nareas columns (8 x 7).}

\item{Smin}{Minimum size limit (mm). Positive real number. Can be a single number (all fleets) or a vector nfleets long (8).}

\item{Smax}{Maximum size limit (mm). Positive real number larger than Smin (if specified). Can be a single number (all fleets) or a vector nfleets long (8).}

\item{BL}{Bag limit (fish retained per trip). Positive integer. Can be a single number (all fleets) or a vector nfleets long (8). Note that predictive models are only available for RecS and HireS in regions FLK and NCFL.}

\item{PRM}{Post Release Mortality (discard mortality). Fraction. Default from US hook and line estimates of Rudershausen et al 2019 (DOI: 10.1002/nafm.10348). Can be a single number or a vector of length nareas (7).}
}
\description{
A function for emulating status quo mahi management and allowing for a wide range of combined management controls. Fleets (n=8) are: (1) USCom, (2) RecN, (3) RecS, (4) HireN, (5) HireS, (6) Intl, (7) Disc, (8) UnRep. Areas (n = 7) are:  (1) NED, (2) NE, (3) NC, (4) SE, (5) SFL, (6) SAR, (7) CAR.
}
\examples{
mahiMP(1, Example_data)
}
\author{
T. Carruthers
}
