% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahiplot.R
\name{mahiplot}
\alias{mahiplot}
\title{Plot overall yield and biomass projections across MPs}
\usage{
mahiplot(
  MSE,
  iqr = 0.8,
  MPcols = c("red", "green", "blue", "orange", "grey", "purple"),
  npy = 20
)
}
\arguments{
\item{MSE}{A class of object 'MMSE' multiMSE}

\item{iqr}{Positive fraction, the interquartile range - default is 0.8 - intervals shown are the 10 percent and 90 percent quantiles of simulations}

\item{MPcols}{Character vector of colors for plotting MP projections}

\item{npy}{Positive integer, number of plot years - No. historical years to plot before projection}
}
\description{
Plot overall yield and biomass projections across MPs
}
\examples{
mahiplot(myMSE)
}
\author{
T. Carruthers
}
